Renames. New cuda memory allocation interface.

+++ src/simulation/HelmholtzSet.cu
@@ -7,7 +7,7 @@
 #include "vector_types.h"
 #include "cuda_float3_operators.cuh"
 #include "cuda_float2_operators.cuh"
-#include "utils.cuh"
+#include "Utils.cuh"
 
 
 constexpr float c_MU0 = 12.566370614e-17F;
@@ -17,8 +17,8 @@ __host__ __device__ HelmholtzSet::HelmholtzSet(unsigned int shape, float sideLen
                                                _sideLen(sideLength), _wireGauge(wireGauge), _coilCenterDistance(coilsDistance),
                                                _coilsPlane(plane), _turns(turns), _dl(dl){
 
-    _externalAngleIncrement_rad = 2*CUDART_PI_F/shape;
-    _internalAngleIncrement_rad = (CUDART_PI_F * shape - 2*CUDART_PI_F)/shape;
+    _externalAngleIncrement_rad = 2*CUDART_PI_F/(float)shape;
+    _internalAngleIncrement_rad = (CUDART_PI_F * (float)shape - 2*CUDART_PI_F)/(float)shape;
 
 }
